refactor(email-worker): move local SQS scripts to infra/local/queue, add retry logic
- Extract sqs-client, setup-queues, local-runner into infra/local/queue
- Add retry loop to setup-queues for slow LocalStack startup
- Update dev script to create queues before API start
- Update dev:email-worker path to infra/local/queue
