+++ infra/local/queue/local-runner.ts
@@ -1,0 +1,62 @@
+import 'dotenv/config';
+
+import {
+  DeleteMessageCommand,
+  ReceiveMessageCommand,
+} from '@aws-sdk/client-sqs';
+import type { Context } from 'aws-lambda';
+
+import { handler } from '../../../src/workers/email/consumer';
+import { createSqsClient, ensureQueues, QUEUE_URL } from './sqs-client';
+
+async function poll(): Promise<void> {
+  const sqs = createSqsClient();
+
+  await ensureQueues(sqs);
+
+  while (true) {
+    try {
+      const response = await sqs.send(
+        new ReceiveMessageCommand({
+          QueueUrl: QUEUE_URL,
+          MaxNumberOfMessages: 10,
+          WaitTimeSeconds: 20,
+        }),
+      );
+
+      if (!response.Messages || response.Messages.length === 0) {
+        continue;
+      }
+
+      const event = {
+        Records: response.Messages.map((msg) => ({
+          messageId: msg.MessageId!,
+          body: msg.Body!,
+          receiptHandle: msg.ReceiptHandle!,
+        })),
+      } as Parameters<typeof handler>[0];
+
+      const result = await handler(event, {} as Context, () => {});
+      const batch = result ?? { batchItemFailures: [] };
+
+      const failedIds = new Set(
+        batch.batchItemFailures.map((f) => f.itemIdentifier),
+      );
+
+      for (const msg of response.Messages) {
+        if (!failedIds.has(msg.MessageId!)) {
+          await sqs.send(
+            new DeleteMessageCommand({
+              QueueUrl: QUEUE_URL,
+              ReceiptHandle: msg.ReceiptHandle!,
+            }),
+          );
+        }
+      }
+    } catch (err) {
+      console.error('Poll error:', err);
+    }
+  }
+}
+
+void poll();
